requirements.txt
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
asttokens==3.0.0
attrs==25.3.0
boto3==1.40.40
botocore==1.40.18
certifi==2025.8.3
charset-normalizer==3.4.3
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
frozenlist==1.7.0
fsspec==2025.9.0
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
multidict==6.6.4
nest-asyncio==1.6.0
numpy==2.3.3
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==21.0.0
Pygments==2.19.2
pyspark==4.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.1.0
requests==2.32.5
s3fs==2025.9.0
s3transfer==0.14.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
wrapt==1.17.3
yarl==1.20.1
